// frontend/lib/types/products.ts
export interface CatalogFilterOption {
  value: string;
  count: number;
}